third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/computed/has.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap.has tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#check-if-stylepropertymap-has-a-property">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="target" style="width: 10px; --foo: auto; background-image: url('A'), url('B')"></div>
<script>
'use strict';

const styleMap = document.getElementById('target').computedStyleMap();

test(() => {
  assert_throws(new TypeError(), () => styleMap.has('lemon'));
}, 'Calling StylePropertyMap.has with an unsupported property throws a TypeError');

const gTestCases = [
  { property: '--Foo', expected: false, desc: 'a custom property not in the property model' },
  { property: 'width', expected: true, desc: 'a valid property' },
  { property: 'wIdTh', expected: true, desc: 'a valid property in mixed case' },
  { property: '--foo', expected: true, desc: 'a valid custom property' },
  { property: 'background-image', expected: true, desc: 'a valid list-valued property' },
];

for (const {property, expected, desc} of gTestCases) {
  test(() => {
    assert_equals(styleMap.has(property), expected);
  }, 'Calling StylePropertyMap.has with ' + desc + ' returns ' + expected);
}

</script>
